Update tests for new write concern behavior
Most of these tests were written at a time when the database was required, so write concern was explicitly enabled. Because the database isn't necessary anymore, the default behavior can be used.

Plus new tests are being added to make sure that write concern is respected by methods that use it.

References #4