Refactor database reads
This update is much in the same vein as cad46aa.

Adding the `_find()` method to the `Model` class will allow for more consistent changes to be made to `find()` and `get()`.

An important change is being made to the API of `find()` and `get()`. The `qs` argument is being deprecated in favor of the new `q` argument. Both arguments will be supported for the time being, but using `qs` will result in a `DeprecationWarning`.