Always set created on new documents
`created` is being given the same treatment as `modified`. Originally the `modified` field was updated by `save()` regardless of any preexisting value. `created`, on the other hand, only updated itself when no value was present (during an insert). Now `created` will automatically be set during an insert. Just as with `modified`, if this behavior isn't desired, `auto_timestamp` should be set to `False`.

Some related comments are being updated as well.

This is also my first attempt to updating the change log in conjunction with the change itself.