Simplify test_save()
`test_save()` is, as its name implies, a test of the `save()` method. Its goal is to make sure that `save()` calls `_update()` correctly. The pieces of the test that deal with `created` and `modified` are being removed so that the test can focus on what it needs to. `created` and `modified` are being covered in `test_save_timestamps()`.