Only update timestamps on successful saves
When using `save()`, the `created` and `modified` fields can be set to new values. The fields should only be updated when the save is successful. This can be done by assigning the values to the internal document after the save.

This is made possible by catching any exception raised by `_update()` and raising it again. Why is this done? Because the assignment to the internal document happens in the try's `else` block.