Remove unnecessary trip to the database when using save_fields()
`Model._update()` uses `find_one()` to reload changed fields when performing an atomic update. With `save_fields()` this is unnecessary because the values are coming from the internal document. Unlike the other atomic update methods, `save_fields()` sets `use_internal` to `True`. Because `use_internal` indicates that the values are already in the internal document, the call to `find_one()` can be skipped.