Add $elemMatch operator
At this point, support is basic. The argument ending with `__elemMatch` takes its query as a dictionary. Any fields used in the dictionary must already be mapped and all operators must already be included.

Eventually I'd like this to feel Pythonic. That could come in the form of supporting `elem_match` in argument names or a method that supports using keyword arguments to construct the query dictionary.

Closes #1