Add a method to perform pop operations
This new method can be used to perform atomic $pop operations. Values can be popped from either end of a list. Specifying the field name (e.g., `field`) will pop a value from the end of the list. Placing a `-` in front of the field name (e.g., `-field`) will pop a value from the beginning of the list.

Closes #13