Move Meta class to separate module
The `Meta` options class is being moved to a separate `meta` module. The decision to do this was made while refactoring `test_meta.py`. I was attempting to clean up the tests and decided that I should target the `Meta` class with many of the tests. Any of the tests related to `ModelMetaClass` are being moved in `test_base.py`.

I feel like I should explain this better, but it has been a long day and I'm just happy to have gotten all the tests in order.

An important thing to note: `TestModelMetaClass.test_subclassed()` was failing on the `collection` option with the classes defined directly in the method. It was also failing with `ModelFactory()`. The test only passes with the classes defined the proper way and at the module level. Because this is how `Model` classes should be used, I'm okay with this.