Update tests for Meta class with proper write concern option
In order to fully deprecate the `safe` option when Simon is used with `MongoClient`, the `Meta` options class needs to set the `w` option instead of the `safe` option. When `safe` is defined, a `DeprecationWarning` should be triggered. These changes to the `TestMeta` should all pass once the necessary changes have been made to `Meta`.

At this point in time, tests should be failing.

References #17