Change Meta class to use write_concern instead of safe or w
This change handles storing a model's default write concern value as `write_concern`. It was originally handled as `safe`. The attribute is needed when calling `delete()`, `insert()`, or `update()`. Having to manage multiple attribute names would get messy. Taking care of it in `Meta` makes it much easier.

Part of this change is adjusting `ModelFactory()` to work with `write_concern` instead of `safe`. In addition, `ModelFactory()` is being updated to hopefully better handle things like what I saw in 68fc36d.

References #17