Pay not attention to the man behind the curtain
This test makes sure that the `Meta` options class triggers a
`DeprecationWarning` for versions of PyMongo with `MongoClient`. Unfortunately
the warning seems to be triggered by another test first, and so
`warnings.catch_warnings()` is unable to catch the warning during
`TestMeta.test_safe_deprecationwarning()`. As a (hopefully temporary)
workaround to this, Mock is being used to mock the warnings module to ensure
the correct warning is being triggered.

References #17