Add typed_fields meta option
This is the first draft of adding a `typed_fields` option to the `Meta` class.
By default `_id` is typed as `ObjectId`. Tests are being added to check that
`Meta.add_to_original()` is handling the attribute correctly.

It currently doesn't actually do anything since `Model` has not yet been made
aware of it.

References #18