Add support for untyped typed fields
Say what?

With this change, `None` can be used as a value in the `typed_fields`
meta option instead of a type to say that the field is untyped. While
this is the behavior when a field isn't added to `typed_fields` at all,
it can be used to say that `_id` should not be typed as an Object ID.

There currently is no code to support this, but it's something I plan to
add.

References #18