Move set_write_concner() to utils
The `base` module is growing increasingly larger. In an effort to trim
it a little, the `set_write_concern` collection of functions is being
moved to the `utils` module. In hindsight, `utils` feels like a more
appropriate home for them anyway.

In addition to moving the functions, a little clean up is being done as
well. The docstrings are being fleshed out to better match the others in
the module. Plus the summary of `set_write_concern_as_w()` is being
fixed as the wording was a bit incorrect. `force_safe` is being renamed
to `force_write_concern` as it's more appropriate given the switch from
`safe` to `w`. Lastly, the `_` is being dropped from the name of each
function.