Adjust warning directives
When RTFD builds the documentation it appears that it doesn't like the
first character of a line of a directive to be a semicolon. It tries to
interpret it as an option of the directive, not as the autodoc directive
I had intended it to be.

The output from RTFD can be seen at
https://readthedocs.org/builds/simon/498201/.