Remove integration tests for PyMongo 2.2 in tox
Now that tox uses the most recent release of each PyMongo release, the
integration tests are getting hung up for PyMongo 2.2. This is the same
as the behavior that was noted for PyMongo 2.3 in b679d6a.

It would appear that this is the result of a call to `getLastError()`.
It was fixed in mongo-python-driver/a33b0fc. More information can be
found in JIRA ticket PYTHON-371:
https://jira.mongodb.org/browse/PYTHON-371