Set minimum version of PyMongo to 2.4
PyMongo 2.4 was a significant milestone as it was the first to include
`MongoClient`. It was released after I began work on Simon, otherwise I
would have used it as the minimum supported version. Since `MongoClient`
is the future, I'd like to simplify my code by dropping support for
`Connection`. This will also mean I can simplify my code by dropping
support for `safe` (`safe` was deprecated in 58d342d).

If someone uses Simon with a version of PyMongo prior to 2.4, a
`DeprecationWarning` will be triggered. This should only happen if
someone installs an older version of PyMongo after they install Simon.

Fixes #21