Improve dictionary iteration
When iterating through the keys of a dictionary it is best to use
something along the lines of `for key in dictionary`. This is the most
efficient way to do it. This won't, however, only you to make changes to
the dictionary's keys. If such behavior is desired, you want to use the
`keys()` or `iterkeys()` methods; they will create a copy of the keys
and allow you to iterate over the copy.

In each of these instances, no changes are being made to the dictionary,
and so the iteration should happen directly on the dictionary, not on a
copy of its keys.