Enforce that field_map is a mapping
If someone specifies a `field_map` that doesn't implmenting `Mapping`, a
`TypeError` will be raised. While this exception is of the right type,
the message associated with it isn't helpful as it applies to the
internal use of `field_map`. This explicit check allows for a meaningful
message with the exception.