Initial commit of the aggregation module
The `aggregation` module is my proposed home for classes related to the
aggregation framework. It may replace `Model.aggregate()` or it may be
utilized by it, I'm not really sure of the end result.

The way I see the module working are that the `Pipeline` class will be
used to construct the pipeline to pass into the `aggregate()` method of
the collection in PyMongo. What is returned will be an instance of some
other class (`ResultSet`?) that will contain the results of the
aggregation query.

At this point all I have is the `project()` method to include and
exclude fields from the result. The `$project` operator is also capable
of renaming fields and constructing embedded fields. When adding that
functionality, I need to keep `field_map` in mind. Fortunately I can use
`$project`'s renaming ability to handle a lot of the work.

References #22