Simplify versioning
When a class is added and given a `versionadded` directive, methods
added as part of the same version don't really need the directive as
well. `versionadded` is being removed from `Pipeline.project()`. Any
other methods added to `Pipeline` under 0.7 will follow this trend as
well.

This is similar to dd0b54a.