Introduce field_map to aggregation
`Pipeline` needs to be aware of what model it is working with so that it
can properly handle that model's `field_map` setting. With this change,
the model can be provided to the `Pipeline` instance as part of its
initialization.

The `project()` method will handle performing the map directly through
the `$project` operator for included fields, and will exclude the
field's real name when for excluded fields.

References #22