Add match() method to Pipeline
The `Pipeline.match()` method can be used to add a `$match` to a
pipeline. It's usage is similar to that of `Model.find()`. Field names
are specified as parameters, and `utils.map_fields()` will take care of
mapping aliases and structuring checks for nested fields.

Just as with `Pipeline.project()`, calls to `match()` can be chained.

References #22