Add support for all lowercase operators
MongoDB operators are case sensitive, using camel case. Simon looks for
operators in the names of keyword arguments. This can lead to code that
doesn't appear pythonic. With this change, `$addToSet` and `$elemMatch`
can now be specified as all lowercase and `utils.map_fields()` will take
care of converting them to the correct case.