Add initial support for Python 3
There is still a lot of work to be done before I'm ready to declare
support for Python 3 complete, but these changes were the minimum
required to get all of the tests to at least pass. Code needs to be
documented better and some parts can probably be refactored to behave
better (e.g., all of the places where `dict.keys()` is wrapped by a call
to `list()` just so I can get the first value).

References #23