Use the correct iterator for dictionary keys
The `iterkeys()` function in the compatibility module is intended to be
used as a substitute for `.keys()` and `.iterkeys()`. As part of the
transition to Python 3, `iterkeys()` should be used instead of these
methods to make sure an iterator is always returned.

Additionally, the `get_next()` function can be used to allow proper
iteration over these iterators. In instances where only the first key is
needed, using `get_next()` is a better approach that using `list()` and
consuming the entire iterator.

References #23