PEP257
Most of the docstrings are being brought up to the guidelines defined in
PEP257. There are some places where the suggested change just didn't
make sense and so the docstring was left alone.

There are also a few places where docstrings are being modified
slightly, usually to make them more consistent with docstrings for
similar classes or functions.